Renamed RightSideMenu to RightMenuDrawer
