+++ app/actions/views/right_menu_drawer.js
