+++ app/scenes/index.js
@@ -6,7 +6,7 @@ import ChannelDrawer from './channel_drawer';
 import ChannelInfo from './channel_info';
 import LoadTeam from './load_team';
 import Login from './login/login_container.js';
-import RightSideMenu from './right_side_menu';
+import RightMenuDrawer from './right_menu_drawer';
 import Root from './root/root_container.js';
 import Search from './search/search_container.js';
 import SelectServer from './select_server/select_server_container.js';
@@ -18,7 +18,7 @@ const scenes = {
     ChannelInfo,
     LoadTeam,
     Login,
-    RightSideMenu,
+    RightMenuDrawer,
     Root,
     Search,
     SelectServer,
