+++ app/scenes/right_menu_drawer/index.js
@@ -1,0 +1,6 @@
+// Copyright (c) 2017 Mattermost, Inc. All Rights Reserved.
+// See License.txt for license information.
+
+import RightMenuDrawerContainer from './right_menu_drawer_container';
+
+export default RightMenuDrawerContainer;
