+++ app/scenes/right_menu_drawer/right_menu_drawer_item.js
