--- app/scenes/right_side_menu/index.js
@@ -1,6 +1,0 @@
-// Copyright (c) 2017 Mattermost, Inc. All Rights Reserved.
-// See License.txt for license information.
-
-import RightSideMenuContainer from './right_side_menu_container';
-
-export default RightSideMenuContainer;
